[SendTabToSelf] Add a STTS receiving infobar on iOS.
Display an infobar on iOS when a tab is received via Sync.
The infobar is only shown if the user has an active and visible
WebState. The other cases will be handled in a future CL.

The infobar includes an open button which opens the tab.

Bug: 944602
Change-Id: I88e593dc77a7301a8b0aa20a8fc74afa1c106270
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1540061
Reviewed-by: Peter Lee <pkl@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: sebsg <sebsg@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#652192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 812855737df6948b2c302263ea0f9caa636d8b5c