Remove old native new tab page controller and associated code.
This removes the kBrowserContainerContainsNTP experiment,
many of the files in c/b/u/ntp, and many of the references to the old
code throughout the codebase.  This includes the added benefit of
finally getting rid of StatusBarHeight().

Bug: 883770, 807330, 826369
Change-Id: Ic5364c243331ede4d4bb8723432319e4d95fe9a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1573176
Commit-Queue: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#652749}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4ee58c972d64d7a36dd2326b9b1f664ad616f5e9