[iOS] Create interface for OverlayManager.
This CL adds the public interface for OverlayManager, which adds the
OverlayRequests supplied to AddRequest() to the OverlayRequestQueue
for that WebState.

OverlayManagers are fetched based on their OverlayModality, which
specifies the modality of the UI delegate's view controller
presentation context.

Bug: 941745
Change-Id: Idc71632131908814cabd5f5c4e7a787de13a41e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1555335
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#653395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c27bf9f13b5b4baea144c2e048c37c42ec32b1e5