[iOS][Alert] Make the alert bigger for a11y
I noticed UIAlertController is wider on accessibility category sizes.
With this change the alert view behaves similar to UIAlertController
with one difference: instead of having 3 sizes (270, 350 and 402) it
it will only have 270 and 402, if the screen is not big enough for 402
it will be as big as it can, respecting the margins.

Bug: 951300
Change-Id: I6cbb7d4ed60aa042567c2ee0d715cb5a267cabb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1583703
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#654812}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2111c65356e2e620d49785c3d8773c60b22c5b19