[iOS] Fix crash in the identity chooser when removing last identity
When IdentityChooserMediator is opened the list of identity can decrease
to zero (if the user opens another app, and remove the last account).
The IdentityChooserMediator is changed to accept nil as the selected
identity.

Also, the ChromeSigninViewController needs to update the primary button
when the last identity is removed (to change the title to "ADD ACCOUNT"),
or when the first account is added (to change the title to "YES I'M IN").
This is done using the UnifiedConsentMediator calling its coordinator,
callings its delegate (ChromeSigninViewController instance).

Bug: 941989
Change-Id: Ibcd2c29f5c0eeccbd941c1910bd8036d952494bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1588556
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#655944}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 83544b07995719c1459b65b9b6ac316a5a9ede1c