Switch ios navigation code to using UnescapeBinaryURLComponent().
This CL is part of an issue to clean up uses of UnescapeURLComponent(),
as it's been used for wildly disparate uses, and much of its logic is
only suitable for cases where we want to format a URL for safe display
to the user. We also want to make it leave more weird characters
escaped, but need to eliminate usage of it in other contexts before we
can do that.

UnescapeBinaryURLComponent() unconditionally unescapes all escaped
characters, and should be used instead in most cases, particulars where
URL_SPECIAL_CHARS_EXCEPT_PATH_SEPARATORS is in use. The primary cases
where it shouldn't be used (without additional checks), are cases where
unescaping path separators or funky characters (like %01) is a problem.
In those places, we should generally fail, instead of partially
unescaping a URL.

Bug: 849998
Change-Id: I7316b30656c42a5897929eef1b4316c55f33efb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1592707
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#656558}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d972cd6a272a18357436f6296e249d15fa87a1e2