[iOS] Fixing issue to enter the sync password in advanced sign-in settings
GoogleServicesSettingsCoordinator needs to open
SyncEncryptionPassphraseTableViewController itself instead of going through
the dispatcher.
When GoogleServicesSettingsCoordinator is opened as the advances sign-in
settings, the sync phrase controller has to be opened in the current
navigation controller.
The dispatcher would open it in the settings view controller.

The AdvancedSigninSettingsNavigationController needs to be a subclass of
SettingsNavigationController in order to support
SyncEncryptionPassphraseTableViewController.

Bug: 959150
Change-Id: I8553252bf44213b3b18d1a500c2734eef1d1ebc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1601041
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#657896}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7f4c2360070f974ba88121566e862f31212a2fa