[ios] Splits kInfobarTypePassword into Save and Update.
This is needed since we need to be able to compare against Save and Update
legacy password Infobars. If not we will have one big bucket with no way
to differentiate if it was originally a Save or Update banner.

Bug: 945478
Change-Id: Icf8bb338080d3ee38469733b0d51e9e347865eab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1597691
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#658247}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: def8256f7d883a18ca324f45ba8d4ebcc54babbb