[ios] Fix page placeholder crash on restore
Adding the page placeholder to a view requires that view to be in a
view hierarchy with the Content Area named guide.
Prior to this CL, the placeholder was attempting to be added to the
web view before the web view is visible. This CL adds a check to see
if the web view is visible. If not, the placeholder will be added when
the web view becomes shown.

TBR=eugenebut@chromium.org

Bug: 956077
Change-Id: I436f7e41f856bcbeb19301a96f738a6dbb572f80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1602392
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#658620}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68c82b7ea44ba1bff19f52bfd956be128d944efa