[ios] Fixes a crash when restoring a closed tab from incognito.
If Recent Tabs is invoked from incognito, restored tabs are placed in
the non-OTR WebStateList.  Before this CL, the current non-OTR WebState
was clobbered and replaced, resulting in a crash if the WebStateList was
empty.  This CL forces restoration to happen in a new WebState, which
sidesteps the crash.

BUG=960742
TEST=Restore a tab from incognito, with no non-Incognito tabs open.  The app should not crash, but the app will also not switch to show the newly-restored tab.

Change-Id: Id38e74cb9bfdada160f1fa049d02219a3a9f1f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1605788
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#659139}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a394b9e8687bd14edcf1ae37e7a44b76fbb5020c