[iOS][web] Use existing navigation item for placeholder load on back navigation
In http://crrev.com/c/1554352, the logic for loading error pages on back/
forward navigation was changed to do a full placeholder load instead of directly
loading the error page, to workaround WebKit caching a blank error page.
However, when navigating back to a page which previously loaded
successfully but is no longer accessible (e.g., because of being in
airplane mode), a full placeholder load destroys forward history.

This CL modifies that change to rewrite the current URL to a placeholder
URL using loadHTMLString, rather than doing a full placeholder load. This
preserves the previous fix while also preserving forward history.

The bug that the previous CL fixed was caused by the following sequence of
loads on forward navigation to an inaccessible page:
(1) A successful placeholder load, caused by forward navigation to an item
    whose URL is a placeholder.
(2) An empty loadHTMLString with the actual URL caused by
    ErrorRetryState::kRetryPlaceholderNavigation.
(3) An attempt to reload the actual URL, cause by
    ErrorRetryState::kNavigatingToFailedNavigationItem.
(4) A loadHTMLSring with the actual URL, to load the error page.

In the above sequence, the failed load step (3) would invalidate the
PageCache entry for the current navigation item. At this point, the item
still has a placeholder URL, so the load in (4) is considered to be
a Standard load, causing the existing content to get added to PageCache
for this item. The URL in the PageCache is that of the DocumentLoader
for the existing content (from (2)), so it's the actual URL. After (4),
navigating away from the page (e.g., going Back) doesn't update the
PageCache, since the item doesn't have a PageCache entry. Then, on the next
forward navigation to this item, WebKit loads the cached content from
(2) along with the actual URL, so when we get the didFinishNavigation call
for this load, we're in case
ErrorRetryState::kDisplayingWebErrorForFailedNavigation and can't tell
whether the page was loaded successfully or not, so assume that it was and
leave it up to the user to reload. At this point the blank content from
(2) is displayed.

The previous CL solves this by doing a full placeholder load after (3).
The problem is that if the page had previously loaded successfully before,
this is a Standard load in WebKit (because it's a loadRequest with URL that
doesn't match the current one), so Forward history is destroyed. The test
added in that CL (ErrorPageTest.BackForwardErrorPage) doesn't catch this,
because the page with the error never has a successful load.

This CL rewrites the URL to a placeholder URL using loadHTMLString rather
than loadRequest. This prevents forward history from being destroyed. It
also still fixes the caching bug: In the sequence above, there will now
be a step (3.5) for the loadHTMLString with empty content and the
placeholder URL. This is the content that will be cached by WebKit. So this
item's URL will still be the placeholder URL on futurue visits, meaning
the ErrorRetryStateMachine will no longer wind up in a state where it can't
tell whether the page has loaded successfully; instead, since the URL is
a placeholder, it will wind up in
ErrorRetryState::kNavigatingToFailedNavigationItem.

Bug: 950489
Change-Id: Ib9437578a84e564fc87702e3a344454baae39029
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1610480
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#659550}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f33d728f0f99cea8fa8ce168e663975863a5d29