Passes AuthenticationService object to NTPHomeMediator (2/4)
ContentSuggestionsCoordinator creates an AuthenticationService object
from BrowserState and passes it to NTPHomeMediator.

NTPHomeMediator can use this AuthenticationService to determine if
user is signed in or not and to fetch the user's avatar if user is
signed in.

Bug: 961120
Change-Id: I1935594130295926c76cfa1f469f8647e7dd4b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1612334
Commit-Queue: Peter Lee <pkl@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#660686}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fd890ec228e76c3b21c82eeab079151d92ee1f0