[ios] Adds startup code to EG2 test bundles.
Registers an NSPrincipalClass in the test bundle's Info.plist, which
provides the name of a class that is automatically instantiated when the
test bundle is loaded into the test runner process.  We can use this
class, along with the XCTestObservation protocol, to run startup code
before any tests are run.

In this CL, we copy pak files into the test runner's bundle and
initialize the ResourceBundle at startup, which allows EG2 tests to use
string resources as needed.

BUG=922813

Change-Id: I348ed41be3939aea2a7525205969a538bbaaefa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1618024
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7c09e61b8f3db5b8e38711042a1463b981a04a38