🦁Add G icon for answers with no specific icon
This CL adds the G icon as a fallback icon for answers with no more
specific icon. It also renames the "currency" imageset to "stock" as
that icon is actually used for the stock price suggestion, not the
currency conversion suggestion (which uses the "conversion" icon).

Bug: 963860
Change-Id: I446a07cbab6b1a5ab0a38629947d09e17522499e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1614219
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661324}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2666ab4096ced139497da883264b90a2fb0c7b67