Run bookmarks earl grey tests on the unified consent enabled bot
This CL updates the bookmarks earl grey test to run with unified consent
feature enabled. It also adds the bookmarks earl grey tests to the list
of tests that are run on the unified consent enabled bot.

This CL also fixes the test testSignInPromoWithWarmStateUsingPrimaryButton
that was actually tapping the secondary sign-in button (not the primary
one), and thus was a duplicate of testSignInPromoWithWarmStateUsingSecondaryButton.

Note: There are a significant number of earl grey tests for bookmarks,
so this CL will significantly increase the time it takes to run the tests
on the ios-simulator-full-configs bots. However the unified consent enabled
bot is temporary and the current plan is to remove it once unified
consent feature is enabled by default.

Bug: 960260
Change-Id: I6a998fb29b1dd352c37dcc9ace516a0254b804e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617513
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661347}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 883fd8e41813757e97bebfe517c04c4020e439ef