[ios] Fixes banner position on iPad
The TabStrip pushes the Ombibox View down and we can't rely on its Guide
for the banner presentation.

This is intended to be a temporary solution and crbug.com/964136 has
been created to fix the NamedGuide or create a new one.

Bug: 911864
Change-Id: I50801c85233bf6a34e3beda3991a7bfa760aa80d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1616381
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661390}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0b33c07a91491b22e819e78d8d5f7e038e4b6508