[iOS] Move the NativeContent code out of WebController
This CL moves the NativeCode part of WebController out to
LegacyNativeContentController.
This object exposes an API to interact with the NativeContent. Its
delegate (the WebController) contains the rest of the methods to allow
interactions with the rest of content.

It also moves the ownership of the NativeContent from
WebControllerContainerView to the LegacyNativeContentController.
The WebControllerContainerView can access the NativeContent through
its delegate.

Bug: 956523
Change-Id: Ie92be0afc95c5e77499bb7d62d3dab15c962a3b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1587891
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#661722}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de8b82098438ec6213a1f5922b5ccbf08bd885ea