ContextMenuCoordinator should not use web::ContextMenuParams
ContextMenuCoordinator uses web::ContextMenuParams as a convenience
when most of the fields in ContextMenuParams are not needed by the
coordinator.

This CL cleans it up by expanding the designated initializer for
ContextMenuCoordinator to take title, view, and location and eliminating
the need for passing an entire struct around.

Change-Id: Iff23e785d09d161e01c1b9f159f605e85ddf4c5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1623541
Commit-Queue: Peter Lee <pkl@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662396}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7786871d814e56dbd0ccebd44dd53528ab86498f