Using tab_test_util.h only in chrome_earl_grey.h
This is part of EG1 to EG2 migration, which involves moving EarlGrey code from
app-side helpers into test code. For GREYAsserts, the policy is to take assert-
containing-helpers, modify them to return NSError* with localizedDescription
containing the assert failure (in the event of a failure), and then assert
on the result of the helper on the test side.

Bug: 922813
Change-Id: I4704c5dc6bccd3eebbb5a5e939737ddbb030cbd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1599825
Commit-Queue: Maksym Onufriienko <monufriienko@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662500}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a410b6988151c6df37713ec0a720e6aedad759bc