[iOS] Clip data: URLs on the tail.
When a long URL is displayed, the steady state omnibox clips it from
the head (and shows only the host), to show the most meaningful part
of the url.
For data: URLs the most meaningful part is the "data:" part, so
it makes more sense to clip on the tail.

Bug: 953294
Change-Id: Ib5404cb85f4f155a3f9cf284343d354262858cdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1617822
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662584}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1fac4beb3e30ffeba4e039d8efd9f7ebf114b0d