🦁Show favicons retrieved from cache immediately, preventing flashing
Before, the favicon completion was always called asynchronously,
leading to the original icon appearing for a split second when a
favicon was found in the cache. Now, that completion will be displayed
instantly.

The asynchronicity was there to simplify the FaviconRetriever logic
(the completion was always called asynchronously), but that isn't
actually super important.

Bug: 964918
Change-Id: I30b7d2f8d9fe0ed9324e48eb994ce677398b388b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1626861
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#662595}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a5e8bc4e384e3be5db2dca3c9cdc4670d2efdb2e