[iOS] Improve accessibility of TableView cells
The accessibility of the TableView isn't great. This CL improves it
by increase the contrast of the sublabels.
It also use a shared constant for the font of the sub labels in case
we need to modify it in the future.

Bug: 960264
Change-Id: I962f66082c459ccf9999848e510670d9e12dca11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1627364
Reviewed-by: Huanzhong Huang <huanzhong@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#663766}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8a9c68d65ad4d0665f15a97602b5219857d320ff